Fix panic: Bad field type
According to
https://pkg.go.dev/github.com/go-playground/validator/v10#hdr-Not_Equal

For strings & numbers, ne will ensure that the value is not equal to the
parameter given. For slices, arrays, and maps, validates the number of items.

So it isn't applicable to pointers. And according to
https://pkg.go.dev/github.com/go-playground/validator/v10#hdr-Required

For slices, maps, pointers, interfaces, channels and functions ensures the value
is not nil.

That's exactly what we need here.