New configuration for control and prometheus services
Example:

``` yaml
listen:
  # control socket for zrepl client, like `zrepl signal` or `zrepl status`.
  - unix: "/var/run/zrepl/control"
    # unix_mode: 0o660		  # write perm for group
    control: true

  # Export Prometheus metrics on http://127.0.0.1:8000/metrics
  - addr: "127.0.0.1:8000"
    # tls_cert: "/usr/local/etc/zrepl/cert.pem"
    # tls_key: "/usr/local/etc/zrepl/key.pem"
    metrics: true
```

One of `addr` or `unix` is required or both of them can be configured. One of
`control` or `metrics` is required or both of them can be configured too.
Everything else is optional. For backward compatibility old style configuration
works too.

While I'm here delete `pprof`. I don't need it and users don't need it too,
because profiling is developer's task, not user's.