Restore and refactor original logic of Sender.ListFilesystems with placeholders
Now it get placeholder states concurrently and disabled by default. I'm not sure
I need this feature, but if, it can be enabled in config:

- name: "test-push"
  type: "push"
  send:
    list_placeholders: true

For reference, I removed original logic in 9b91e31 with message:

It was getting placeholder state for every dataset, but it uses filter and
filter returns only non placeholder datasets. So every returned fs isn't
placeholder and it was exec `zfs get` for every datasets without reason.